+++ libs/app/directives/code-highlighter/code-highlighter.directive.ts
@@ -1,37 +1,17 @@
-import { Directive, ElementRef, HostBinding, Input, OnChanges } from '@angular/core';
-import { HighlightResult } from 'highlight.js';
-import highlight from 'highlight.js/lib/core';
-import xml from 'highlight.js/lib/languages/xml';
-
-highlight.registerLanguage('html', xml);
-highlight.registerLanguage('xml', xml);
+import { Directive, ElementRef, inject, input, OnChanges } from '@angular/core';
+import { NgDocHighlighterService } from '@ng-doc/app/services';
 
 @Directive({
-	selector: 'code[ngDocCodeHighlighter]',
-	standalone: true,
+  selector: '[ngDocHighlighter]',
+  standalone: true,
 })
 export class NgDocCodeHighlighterDirective implements OnChanges {
-	@Input('ngDocCodeHighlighter')
-	code: string = '';
+  code = input.required<string>({ alias: 'ngDocHighlighter' });
 
-	@Input()
-	html: string = '';
+  protected readonly element = inject(ElementRef<HTMLElement>).nativeElement;
+  protected readonly highlighter = inject(NgDocHighlighterService);
 
-	@Input()
-	language: string = 'typescript';
-
-	@HostBinding('class.hljs')
-	protected highlightJsClass: boolean = true;
-
-	constructor(private readonly elementRef: ElementRef<HTMLElement>) {}
-
-	ngOnChanges(): void {
-		if (this.code) {
-			const result: HighlightResult = highlight.highlight(this.code, { language: this.language });
-
-			this.elementRef.nativeElement.innerHTML = result.value ?? this.html;
-		} else {
-			this.elementRef.nativeElement.innerHTML = this.html;
-		}
-	}
+  ngOnChanges(): void {
+    this.element.innerHTML = this.highlighter.highlight(this.code());
+  }
 }