+++ libs/app/providers/ng-doc-app/provide-ng-doc-app.ts
@@ -1,5 +1,6 @@
 import { ViewportScroller } from '@angular/common';
-import { APP_INITIALIZER, Provider } from '@angular/core';
+import { APP_INITIALIZER, inject, Provider } from '@angular/core';
+import { NgDocHighlighterConfig, NgDocHighlighterService } from '@ng-doc/app/services/highlighter';
 import { NgDocUiConfig, provideNgDocUiKitConfig } from '@ng-doc/ui-kit';
 
 /**
@@ -10,6 +11,10 @@ export interface NgDocApplicationConfig {
    * UI Kit configuration.
    */
   uiKit?: NgDocUiConfig;
+  /**
+   * Shiki theme.
+   */
+  shiki?: NgDocHighlighterConfig;
 }
 
 /**
@@ -25,6 +30,16 @@ export function provideNgDocApp(config?: NgDocApplicationConfig): Provider[] {
       deps: [ViewportScroller],
       useFactory: (viewportScroller: ViewportScroller) => {
         return () => viewportScroller.setOffset([0, 120]);
+      },
+    },
+    /** --- Shiki --- */
+    {
+      provide: APP_INITIALIZER,
+      multi: true,
+      useFactory: () => {
+        const highlighter = inject(NgDocHighlighterService);
+
+        return () => highlighter.initialize(config?.shiki);
       },
     },