+++ libs/app/services/index.ts
@@ -1,2 +1,3 @@
+export * from '@ng-doc/app/services/highlighter';
 export * from '@ng-doc/app/services/sidebar';
 export * from '@ng-doc/app/services/store';