+++ libs/app/tokens/index.ts
@@ -3,5 +3,6 @@ export * from './mermaid';
 export * from './page-processor';
 export * from './page-skeleton';
 export * from './route-prefix';
+export * from './shiki-theme';
 export * from './theme';
 export * from './type-control';