+++ libs/builder/engine/builders/page/demo-assets.builder.ts
@@ -5,8 +5,6 @@ import {
   createMainTrigger,
   NgDocBuilderContext,
   NgDocComponentAsset,
-  postProcessHtml,
-  processHtml,
   renderTemplate,
   replaceKeywords,
   runBuild,
@@ -16,7 +14,7 @@ import * as path from 'path';
 import { of } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
-import { getDemoAssets, getDemoClassDeclarations } from '../../../helpers';
+import { getDemoAssets, getDemoClassDeclarations, UTILS } from '../../../helpers';
 import { Builder, watchFile } from '../../core';
 import { EntryMetadata } from '../interfaces';
 
@@ -70,13 +68,16 @@ export function demoAssetsBuilder(config: Config): Builder<AsyncFileOutput> {
 
         for (const [, value] of Object.entries(demoAssets)) {
           for (const asset of value) {
-            const processed = await processHtml(asset.code, {});
+            const processed = await UTILS.processHtml(asset.code, {
+              lightTheme: config.context.config.shiki?.themes.light,
+              darkTheme: config.context.config.shiki?.themes.dark,
+            });
 
             if (processed.error) {
               throw processed.error;
             }
 
-            const postProcessed = await postProcessHtml(processed.content);
+            const postProcessed = await UTILS.postProcessHtml(processed.content);
 
             if (postProcessed.error) {
               throw postProcessed.error;