+++ libs/builder/engine/core/operators/global-trigger.ts
@@ -1,7 +1,0 @@
-/**
- *
- * @param id
- */
-export function registerGlobalTrigger(id: string): void {
-	console.log('registerGlobalTrigger', id);
-}