+++ libs/builder/engine/entities/plugins/types/entity-plugin.ts
@@ -1,6 +1,0 @@
-import { NgDocEntity } from '../../abstractions/entity';
-
-export interface NgDocEntityPlugin<T, TEntity extends NgDocEntity = NgDocEntity> {
-	id: string;
-	execute: (data: T, context: TEntity) => Promise<T>;
-}