+++ libs/utils/html/plugins/highlight-code-lines.ts
@@ -1,4 +1,4 @@
-import { Element, ElementContent, Root } from 'hast';
+import { Element, Root } from 'hast';
 import { visit } from 'unist-util-visit';
 
 import { attrValue, isCodeBlock, isElement } from '../helpers';
@@ -9,19 +9,19 @@ import { attrValue, isCodeBlock, isElement } from '../helpers';
 export default function highlightCodeLines() {
   return (tree: Root) => {
     visit(tree, 'element', (node: Element, _i, parent: Root | Element | null) => {
-      if (isCodeBlock(parent, node)) {
-        const highlightedLines: number[] = JSON.parse(attrValue(node, 'highlightedLines') ?? '[]');
+      if (isCodeBlock(parent, node) && parent?.type === 'element') {
+        const highlightedLines: number[] = JSON.parse(
+          attrValue(parent, 'highlightedLines') ?? '[]',
+        );
 
         highlightedLines.length &&
-          node.children.forEach((child: ElementContent, i: number) => {
+          node.children.filter(isElement).forEach((child: Element, i: number) => {
             // eslint-disable-next-line @typescript-eslint/ban-ts-comment
             // @ts-ignore
-            if (isElement(child) && child.properties.class?.includes('line')) {
-              if (highlightedLines.includes(i + 1)) {
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-ignore
-                child.properties.class.push('highlighted');
-              }
+            if (child.properties.class === 'line' && highlightedLines.includes(i + 1)) {
+              // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+              // @ts-ignore
+              child.properties.class = 'line highlighted';
             }
           });
       }