+++ libs/utils/html/replace-keywords.ts
@@ -3,7 +3,6 @@ import rehypeStringify from 'rehype-stringify';
 import { unified, VFileWithOutput } from 'unified';
 
 import keywordsPlugin, { GetKeyword } from './plugins/keywords.plugin';
-import markCodeBlocksPlugin from './plugins/mark-code-blocks.plugin';
 
 export interface ReplaceKeywordsConfig {
   getKeyword: GetKeyword;
@@ -24,7 +23,6 @@ export async function replaceKeywords(
     .use(rehypeParse, { fragment: true })
     .use(rehypeStringify)
     .use(keywordsPlugin, { getKeyword })
-    .use(markCodeBlocksPlugin)
     .process(html)
     .then((file: VFileWithOutput<string>) => file.toString());
 }