+++ libs/utils/index.ts
@@ -1,4 +1,5 @@
-export * from './html/html-processor';
 export * from './html/post-process-html';
+export * from './html/process-html';
 export * from './html/remove-not-indexable-content';
 export * from './html/replace-keywords';
+export * from './html/stringify-entities';